moved conditional two moons to tests
