--- lbi/models/flows/__init__.py
@@ -1,2 +1,0 @@
-from .maf import MaskedAffineFlow
-from .base import InitializeFlow
